Reorganise several data model tests into new file test_01.py
